+++ src/main/kotlin/com/lgguan/iot/position/entity/AoaDataInfo.kt
@@ -22,9 +22,9 @@ class AoaDataInfo: Model<AoaDataInfo>() {
     var mapId: String? = null
     var zoneId: String? = null
     var optScale: Float? = null
-    var posX: Float? = null
-    var posY: Float? = null
-    var posZ: Float? = null
+    var posX: Double? = null
+    var posY: Double? = null
+    var posZ: Double? = null
     var timestamp: Long? = null
     var createTime: Date? = null
     var status:Int? = null  //0:freezing 1:moving
